Map update (v140)
- Apparently game/client crashes on server join when there is a
"func_breakable_surf" object in game. Changed it to: "func_breakable"
and map is back to normal
- After testing the "func_breakable" glass object, it still needs some
more refinements in settings of when to break. Right now it breaks when
you walk into it. Also, need to add sounds of breaking glass. The glass
is hard to see, so perhaps find a better texture for glass?
- Main entrance doors are slightly too low and get confused which one to
open once one door is open
- Need to update main entrance door sounds as they aren't working
properly (set to only sound on closing door)
- Need to refine bushes in front of garage and to the side as they still
look kind of strange in game